able to take input too
