+++ src/TodoShow.js
@@ -4,18 +4,18 @@ import ExpansionPanel from '@material-ui/core/ExpansionPanel';
 import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
 import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import Content from './Content';
-class Completed extends Component{
+
+class TodoShow extends Component{
 render(){
     return (
         <div>
 
         <ExpansionPanel>
         <ExpansionPanelSummary expandIcon={<ExpandMoreIcon />}>
-          <Typography variant="title">Completed</Typography>
+          <Typography variant="title">To-do List here</Typography>
         </ExpansionPanelSummary>
         <ExpansionPanelDetails>
-        <Content />
+      
         </ExpansionPanelDetails>
       </ExpansionPanel>
 
@@ -25,4 +25,4 @@ render(){
 }
     //it should be a drop down menu
 }
-export default Completed;
+export default TodoShow;
